continued gui example, supporting Windows for development and validation.

+++ examples/gui/run.sh
@@ -1,9 +1,20 @@
-export PATH=$PATH:~/.platformio/penv/bin/
-rm -rf .pio
-# compile and upload firmware to target
-pio run -t upload
-# compile and run receiver application
-# you want to install sdl2 and libserialport via Homebrew on mac. use choco on Windows.
-cmake -S ./renderer/ -B ./renderer/build
-cmake --build ./renderer/build
-./renderer/build/tinyrtos_gui_renderer
+MODE="both"
+
+if [ "$1" = "-r" ]; then
+  MODE="renderer"
+elif [ "$1" = "-f" ]; then
+  MODE="firmware"
+fi
+
+export PATH="$PATH:$HOME/.platformio/penv/bin"
+
+if [ "$MODE" = "firmware" ] || [ "$MODE" = "both" ]; then
+  rm -rf .pio
+  pio run -t upload
+fi
+
+if [ "$MODE" = "renderer" ] || [ "$MODE" = "both" ]; then
+  cmake -S ./renderer/ -B ./renderer/build
+  cmake --build ./renderer/build
+  ./renderer/build/tinyrtos_gui_renderer
+fi
